feat: reduce chain config updates
